requirements.txt
streamlit>=1.53.0
openpyxl>=3.1.5
requests>=2.32.0
numpy>=2.2.0
python-dotenv>=1.0.1
pytest>=9.0.0
